Move fs/mod.go to mod/mod.go
